# Add support for CompletionList.CommitCharacters.