- `CompletionList.CommitCharacters` serves as a way for completion items to not redundantly provide commit characters. Aka, all completion list commit characters serve as a default for underlying completion items.
- The LSP Platform hasn't yet inserted these changes into VS; however, our language server doesn't depend on the LSP platform directly anyhow. This changeset enables the commit character optimization of the LSP platform in a way that will automatically light up once it becomes available.
- Updated benchmarks to use new CommitCharacters on the completion list.
- Expanded our language server client capabilities to track VS' completion list capabilities that identify whether or not the CommitCharacters. This manifests as `PlatformAgnosticCompletionCapability` and `VSCompletionListCapability`.
 - Added a new `VSCompletionList` and `OptimizedVSCompletionList` set of types with their optimized forms.

**Before:**
|                                              Method |      Mean |     Error |    StdDev |   Op/s |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|---------:|---------:|--------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 12.302 ms | 0.2425 ms | 0.4786 ms |  81.29 | 718.7500 | 359.3750 |       - |  4,569 KB |
|           'Component Completion List Serialization' |  2.152 ms | 0.0319 ms | 0.0283 ms | 464.75 |  82.0313 |  31.2500 | 15.6250 |    555 KB |
|         'Component Completion List Deserialization' |  9.409 ms | 0.1797 ms | 0.2399 ms | 106.28 | 640.6250 | 312.5000 |       - |  4,013 KB |

**After:**
|                                              Method |     Mean |     Error |    StdDev |  Op/s |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|---------------------------------------------------- |---------:|----------:|----------:|------:|---------:|---------:|--------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 7.434 ms | 0.1169 ms | 0.1036 ms | 134.5 | 585.9375 | 296.8750 |  7.8125 |  3,679 KB |
|           'Component Completion List Serialization' | 1.924 ms | 0.0257 ms | 0.0241 ms | 519.8 |  83.9844 |  37.1094 | 21.4844 |    523 KB |
|         'Component Completion List Deserialization' | 5.888 ms | 0.1090 ms | 0.1908 ms | 169.8 | 507.8125 | 242.1875 |       - |  3,156 KB |

**Results:**
Roundtrip: 1.7x faster
Serialization: 1.1x faster
Deserialization: 1.6x faster

Razor's part of dotnet/aspnetcore#31886